AutoScaleX Pro 2.2/requirements.txt
requests>=2.31.0,<3.0
python-telegram-bot>=20.0,<22.0
python-dotenv>=1.0.0,<2.0
cryptography>=41.0.0
pytest>=7.0
pytest-asyncio>=0.21.0
